HTBHF-2731 refactor decision handler constants
