+++ src/web/routes/application/steps/common/constants.js
@@ -1,6 +1,5 @@
 module.exports.YES = 'yes'
 module.exports.NO = 'no'
-module.exports.ELIGIBLE = 'ELIGIBLE'
 module.exports.EMAIL = 'email'
 module.exports.TEXT = 'text'
 module.exports.CONFIRMATION_CODE_SESSION_PROPERTY = 'confirmationCode'
