+++ src/web/routes/application/steps/decision/decision.js
@@ -2,7 +2,7 @@ const { configureSessionDetails, handleRequestForPath } = require('../../flow-co
 const { DECISION_URL, prefixPath } = require('../../paths')
 const { isUndefined } = require('../../../../../common/predicates')
 const { getDecisionStatus } = require('./get-decision-status')
-const { FAIL } = require('./statuses')
+const { FAIL } = require('./decision-statuses')
 const { getDecisionPageFallback } = require('./decision-fallback')
 
 const STATUS_TEMPLATE_MAP = {
