+++ src/web/routes/application/steps/decision/eligibility-statuses.js
@@ -1,0 +1,1 @@
+module.exports.ELIGIBLE = 'ELIGIBLE'
